Non-intrusive option to select one of the 3 virtualization technologies supported in Hydrogen release : opendove, ovsdb, vtn.
The basic idea is to make use of the Apache Felix FileInstall to filter out the excluded technology.
In order to avoid introducing multiple config files and keep the changes contained in the Integration branch & Virtualization edition,
we decided to play with a maven assembly plugin option which essentially provides a way to wrap the root run.sh with a virtualization
edition's run.sh.

1. The root run.sh from the base edition is renamed to run.base.sh
2. Introduced a new run.sh (in assemble/resources folder - Note I didnt add it in main/resources to avoid conflicts)
   which is essentially a wrapper for run.base.sh.
3. Maven Assembly plugin takes care of the rest with the <file> and <fileSet> changes.

Change-Id: I76554e60fd3360f6ceea2fec06c98897cac4b99b
Signed-off-by: Madhu Venugopal <mavenugo@gmail.com>