virtualzation run.sh: launch run.base.sh from same path as run.sh
The virtualization edition's run.sh script internally hard codes
the underlying run.base.sh script to be in cwd:

  ./run.base.sh ...

Use a trick similar to the base edition's run.sh script to redirect to
the underlying [run.base.sh] script to use the same path as run.sh.

Note: A platform (like Linux) is likely to place a top-level run.sh
script in a /bin type directory and the underlying internal scripts in
an out-of-$PATH directory like /usr/libexec/.  This may need some final
update based on that, but for now brings the two editions' scripts in
line with one another.

Change-Id: Ia8d719826de02efa6a36d3427f880cf062edc547
Signed-off-by: Chris Wright <chrisw@sous-sol.org>