Add bgpcep to rpm packaging.
This is just a placeholder for bgpcep. The spec file is mostly complete, but
there is an issue in that bgpcep does not keep the artifacts within the build
directory and only pushes them to the maven repo. Thus, the artifacts cannot
be properly packaged. Committing the file so when that issue is fixed the
packaging will be ready.

Change-Id: If87952b94865d783c9b0717524f7a04a78b0628c
Signed-off-by: Sam Hague <shague@redhat.com>