Bug 1099: Remove toaster from virt edition
In the virt edition, we use a clever felix dm
hack to allow the user to pick one or the other
virt providers by filtering out potentially
conflicting bundles.

vtn has chosen to filter out all sample apps
so as to avoid collisions with things like
simpleforwarder.

Toaster is a sample app and thus its bundle
is filtered.  However, there is configuration
present that demands toaster in order for it
to be applied.  This causes Bug 1099.

The simplest way to resolve this, rather than
extending the Felix DM hack, is to filter
toaster and its config file out of
the virtualization edition.

I have verified that

./run.sh -virt vtn -of13

starts the controller without the reported error.

Change-Id: I0f45ef4f183b3458bbc4b3970e21e833d1cdaccf
Signed-off-by: Ed Warnicke <eaw@cisco.com>