Logically separate OF plugin from controller distribution bundle.
This plugin effectively results in no changes in output, but introduces
the correct logical separation of the OFplugin from controller.

When copying the controller distribution.opendaylight and unpacking
it, the legacy OF plugin is excluded at the file level.

However, then it is added in as a proper maven *dependency*.

This way, OFplugin default changes can be independent of whatever
is happening in controller opendaylight.distribution.

Change-Id: I4ae8167432eb0632a6768a05037daa0b8f630247
Signed-off-by: Ed Warnicke <eaw@cisco.com>