Install ODL on Vagrant base box via Packer+Ansible
Use the ODL Ansible role to install OpenDaylight on
the ODL Vagrant base box via the ansible-local Packer
provisioner.

All documentation updates included.

Attempted to clean up the VM as much as possible, to
minimize the size of the resulting base box. I think
there are still optimizations to be made, but most of
the size is consumed by required things like ODL, Java,
Ansible (for ansible-local) and their dependences.

Change-Id: If11d2308036b44feb5bada86482b7e4413468735
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>